requirements.txt
google-genai
fastapi
pydantic
langchain-google-genai
logging
mcp
mcp[cli]
json